cambie la variable Monto de int a double

+++ app/src/main/java/edu/ucne/jose_estrella_ap2_p2/data/remote/dto/GastoDto.kt
@@ -8,7 +8,7 @@ class GastoDto (
     val suplidor : String,
     val ncf: String,
     val itbis: Double,
-    val monto: Int
+    val monto: Double
 )
 {
     fun toDomain() = Gasto(gastoId,fecha,suplidor,ncf,itbis,monto)
